+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - The `ResponseBuilder` now requires PSR-17 factories instead of a PHP-HTTP factory, to align it with the `Client`. This is only a breaking change in the rare case where you provide your own factory to the `ResponseBuilder`.
 - Removed `domainAliases` from constructor arguments. Please use `setDomainAliases()` on the instance instead.
 
+### Removed
+- Dropped PHP <7.4 support.
+
 ## [2.3.2] - 2021-08-30
 
 ### Fixed